BackEnd/config/chatbot/requirements.txt
ollama
pytesseract
pillow
pdf2image
python-docx
PyPDF2
opencv-python
numpy
regex
matplotlib
argparse